edits

+++ src/main/java/com/Akansh/Sampark/Repository/UserDAO.java
@@ -2,10 +2,16 @@ package com.Akansh.Sampark.Repository;
 
 
 import com.Akansh.Sampark.Models.User;
+import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface UserDAO extends JpaRepository<User, Integer> {
+public interface UserDAO extends JpaRepository<User, Long> {
+
+
 
 }
